LeDucAnh_v3_v17
LeDucAnh_v3_v17

+++ Repository/ViewModels/CategoryDetailsViewModel.cs
@@ -17,7 +17,7 @@ namespace Repository.ViewModels
 
         public List<StoreDetailsViewModels> StoreDetailViewModel { get; set; }
         public int Number { get; set; } = 0;
-        public decimal Commission { get; set; }
+        public float Commission { get; set; }
         public DateTime? CreatedDate { get; set; }
         public DateTime? ModifiedDate { get; set; }
     }
